pyproject.toml
[tool.poetry]
name = "flysight"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.api.dependencies]
fastapi = {extras = ["standard"], version = "^0.112.2"}
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.32"
uvicorn = "^0.30.6"
influxdb-client = "^1.45.0"
pandas = "^2.2.2"
numpy = "^2.1.0"
geopy = "^2.4.1"
peakutils = "^1.3.5"
pyproj = "^3.6.1"
python-dotenv = "^1.0.1"
utm = "^0.7.0"
bcrypt = "^4.2.0"
python-multipart = "^0.0.9"
loguru = "^0.7.2"
pydantic = {extras = ["email"], version = "^2.8.2"}


[tool.poetry.group.notebook.dependencies]
numpy = "^2.1.0"
pandas = "^2.2.2"
plotly = "^5.23.0"
peakutils = "^1.3.5"
geopy = "^2.4.1"
pyproj = "^3.6.1"
utm = "^0.7.0"
seaborn = "^0.13.2"
pyarrow = "^17.0.0"
python-dotenv = "^1.0.1"
panel = "^1.4.5"
ipywidgets = "^8.1.5"
import-ipynb = "^0.1.4"


[tool.poetry.group.scraper.dependencies]
selenium = "^4.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
